renamed routes/users.py to routes/users_routes.py. Removed rendertemplate from app.py
